Revert "feat!: rename `@tutorialkit/components-react` to `@tutorialkit/react` (#155)"
This reverts commit e3c0fee902a7bfc312fb01b30531209815d460c3.
