+++ docs/tutorialkit.dev/src/components/react-examples/Example.astro
@@ -1,5 +1,5 @@
 ---
-import { classNames } from '@tutorialkit/react';
+import { classNames } from '@tutorialkit/components-react';
 import '@tutorialkit/astro/default-theme.css';
 
 interface Props {
