+++ packages/astro/src/default/components/ResizablePanel.astro
@@ -1,5 +1,5 @@
 ---
-import { classNames } from '@tutorialkit/react';
+import { classNames } from '@tutorialkit/components-react';
 
 export type Type = 'horizontal' | 'vertical';
 export type Priority = 'min' | 'max';
