+++ packages/astro/src/index.ts
@@ -87,14 +87,14 @@ export default function createPlugin({
           vite: {
             optimizeDeps: {
               entries: ['!**/src/(content|templates)/**'],
-              include: process.env.TUTORIALKIT_DEV ? [] : ['@tutorialkit/react'],
+              include: process.env.TUTORIALKIT_DEV ? [] : ['@tutorialkit/components-react'],
             },
             define: {
               __ENTERPRISE__: `${!!enterprise}`,
               __WC_CONFIG__: enterprise ? JSON.stringify(enterprise) : 'undefined',
             },
             ssr: {
-              noExternal: ['@tutorialkit/astro', '@tutorialkit/react'],
+              noExternal: ['@tutorialkit/astro', '@tutorialkit/components-react'],
             },
             plugins: [
               userlandCSS,
