+++ packages/components/react/scripts/build.js
