+++ packages/components/react/src/BootScreen.tsx
