+++ packages/components/react/src/Panels/TerminalPanel.tsx
