+++ packages/components/react/src/core/CodeMirrorEditor/BinaryContent.tsx
