+++ packages/components/react/src/core/CodeMirrorEditor/themes/vscode-dark.ts
