+++ packages/components/react/src/core/Terminal/index.tsx
