+++ packages/components/react/src/core/Terminal/theme.ts
