+++ packages/components/react/src/styles/resize-panel.module.css
