+++ packages/components/react/src/styles/terminal.css
